модуль редиса перенесен в базы данных прошелся блеком
