+++ app/core/security.py
@@ -5,7 +5,7 @@ import logging
 from jose import jwt
 from passlib.context import CryptContext
 from fastapi import Response
-from app.core.redis import RedisSingleton
+from app.core.database import RedisSingleton
 from cryptography.fernet import Fernet
 
 
@@ -16,7 +16,9 @@ class Security:
     def __init__(self):
         self.pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
         self.cipher = Fernet(os.getenv("ENCRYPTION_KEY"))
-        self.ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 300))
+        self.ACCESS_TOKEN_EXPIRE_MINUTES = int(
+            os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 300)
+        )
         self.REFRESH_TOKEN_EXPIRE_DAYS = int(os.getenv("REFRESH_TOKEN_EXPIRE_DAYS", 30))
         self.CSRF_TOKEN_EXPIRE_MINUTES = int(os.getenv("CSRF_TOKEN_EXPIRE_MINUTES", 60))
         self.SECRET_KEY = os.getenv("SECRET_KEY")
